Django_Auth_System/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-allauth==65.4.1
django-htmx==1.22.0
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
